+++ lib/src/appodeal.dart
@@ -347,7 +347,7 @@ class Appodeal {
 
   /// Gets SDK version.
   static String getSDKVersion() {
-    return "3.3.3.1";
+    return "3.4.1";
   }
 
   /// Gets SDK platform version.